pyproject.toml
[project]
name = "uv-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy>=2.0.37",
    "sample",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "alembic>=1.14.0",
    "factory-boy>=3.3.1",
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "ruff>=0.9.1",
]

[tool.uv.sources]
sample = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q -s"
testpaths = [
    "tests",
    "integration",
]
